Created LooseAppSupport file as intermediary between DevMojo and PluginConfigSupport
